Add files via upload
As a novice analyst, I embarked on this project to investigate Airbnb in Seattle, WA. The main idea of the project was to create a visualization in Tableau. However, the data that I downloaded had an error in the date column, and to fix this problem, I decided to use MySQL.

Keep in mind that this project is part of @AlexTheAnalyst's bootcamp. While the bootcamp focused on using SQL Server, I encountered challenges when attempting to write the code for MySQL. This was not an exception as the bootcamp used the data only in Tableau. The mentioned obstacle presented an opportunity for me to explore and discover alternative solutions, enabling me to develop coding skills in different environments. Despite the initial difficulty, I was able to complete the project successfully, utilizing MySQL to correct the error and present my findings in a comprehensive report. The experience allowed me to further my knowledge and provided a valuable opportunity to develop my analytical and problem-solving skills.